+++ tests/download_test.rs
@@ -1,5 +1,7 @@
+#![allow(clippy::disallowed_names)]
+
 use proto_pdk_test_utils::*;
-use starbase_sandbox::{create_empty_sandbox, locate_fixture};
+use starbase_sandbox::locate_fixture;
 use std::collections::HashMap;
 
 generate_download_install_tests!(
@@ -10,12 +12,13 @@ generate_download_install_tests!(
 
 #[test]
 fn supports_linux_arm64() {
-    let sandbox = create_empty_sandbox();
-    let plugin = create_schema_plugin_with_config(
+    let sandbox = create_empty_proto_sandbox();
+    let plugin = sandbox.create_schema_plugin_with_config(
         "schema-test",
-        sandbox.path(),
         locate_fixture("schemas").join("bins.toml"),
-        HashMap::from_iter([map_config_environment(HostOS::Linux, HostArch::Arm64)]),
+        |config| {
+            config.host(HostOS::Linux, HostArch::Arm64);
+        },
     );
 
     assert_eq!(
@@ -38,12 +41,13 @@ fn supports_linux_arm64() {
 
 #[test]
 fn supports_linux_x64() {
-    let sandbox = create_empty_sandbox();
-    let plugin = create_schema_plugin_with_config(
+    let sandbox = create_empty_proto_sandbox();
+    let plugin = sandbox.create_schema_plugin_with_config(
         "schema-test",
-        sandbox.path(),
         locate_fixture("schemas").join("bins.toml"),
-        HashMap::from_iter([map_config_environment(HostOS::Linux, HostArch::X64)]),
+        |config| {
+            config.host(HostOS::Linux, HostArch::X64);
+        },
     );
 
     assert_eq!(
@@ -66,12 +70,13 @@ fn supports_linux_x64() {
 
 #[test]
 fn supports_macos_arm64() {
-    let sandbox = create_empty_sandbox();
-    let plugin = create_schema_plugin_with_config(
+    let sandbox = create_empty_proto_sandbox();
+    let plugin = sandbox.create_schema_plugin_with_config(
         "schema-test",
-        sandbox.path(),
         locate_fixture("schemas").join("bins.toml"),
-        HashMap::from_iter([map_config_environment(HostOS::MacOS, HostArch::Arm64)]),
+        |config| {
+            config.host(HostOS::MacOS, HostArch::Arm64);
+        },
     );
 
     assert_eq!(
@@ -94,12 +99,13 @@ fn supports_macos_arm64() {
 
 #[test]
 fn supports_macos_x64() {
-    let sandbox = create_empty_sandbox();
-    let plugin = create_schema_plugin_with_config(
+    let sandbox = create_empty_proto_sandbox();
+    let plugin = sandbox.create_schema_plugin_with_config(
         "schema-test",
-        sandbox.path(),
         locate_fixture("schemas").join("bins.toml"),
-        HashMap::from_iter([map_config_environment(HostOS::MacOS, HostArch::X64)]),
+        |config| {
+            config.host(HostOS::MacOS, HostArch::X64);
+        },
     );
 
     assert_eq!(
@@ -122,12 +128,13 @@ fn supports_macos_x64() {
 
 #[test]
 fn supports_windows_arm64() {
-    let sandbox = create_empty_sandbox();
-    let plugin = create_schema_plugin_with_config(
+    let sandbox = create_empty_proto_sandbox();
+    let plugin = sandbox.create_schema_plugin_with_config(
         "schema-test",
-        sandbox.path(),
         locate_fixture("schemas").join("bins.toml"),
-        HashMap::from_iter([map_config_environment(HostOS::Windows, HostArch::Arm64)]),
+        |config| {
+            config.host(HostOS::Windows, HostArch::Arm64);
+        },
     );
 
     assert_eq!(
@@ -150,12 +157,13 @@ fn supports_windows_arm64() {
 
 #[test]
 fn supports_windows_x86() {
-    let sandbox = create_empty_sandbox();
-    let plugin = create_schema_plugin_with_config(
+    let sandbox = create_empty_proto_sandbox();
+    let plugin = sandbox.create_schema_plugin_with_config(
         "schema-test",
-        sandbox.path(),
         locate_fixture("schemas").join("bins.toml"),
-        HashMap::from_iter([map_config_environment(HostOS::Windows, HostArch::X86)]),
+        |config| {
+            config.host(HostOS::Windows, HostArch::X86);
+        },
     );
 
     assert_eq!(
@@ -178,12 +186,13 @@ fn supports_windows_x86() {
 
 #[test]
 fn locates_linux_bin() {
-    let sandbox = create_empty_sandbox();
-    let plugin = create_schema_plugin_with_config(
+    let sandbox = create_empty_proto_sandbox();
+    let plugin = sandbox.create_schema_plugin_with_config(
         "schema-test",
-        sandbox.path(),
         locate_fixture("schemas").join("bins.toml"),
-        HashMap::from_iter([map_config_environment(HostOS::Linux, HostArch::Arm64)]),
+        |config| {
+            config.host(HostOS::Linux, HostArch::Arm64);
+        },
     );
 
     assert_eq!(
@@ -203,12 +212,13 @@ fn locates_linux_bin() {
 
 #[test]
 fn locates_macos_bin() {
-    let sandbox = create_empty_sandbox();
-    let plugin = create_schema_plugin_with_config(
+    let sandbox = create_empty_proto_sandbox();
+    let plugin = sandbox.create_schema_plugin_with_config(
         "schema-test",
-        sandbox.path(),
         locate_fixture("schemas").join("bins.toml"),
-        HashMap::from_iter([map_config_environment(HostOS::MacOS, HostArch::X64)]),
+        |config| {
+            config.host(HostOS::MacOS, HostArch::X64);
+        },
     );
 
     assert_eq!(
@@ -228,12 +238,13 @@ fn locates_macos_bin() {
 
 #[test]
 fn locates_windows_bin() {
-    let sandbox = create_empty_sandbox();
-    let plugin = create_schema_plugin_with_config(
+    let sandbox = create_empty_proto_sandbox();
+    let plugin = sandbox.create_schema_plugin_with_config(
         "schema-test",
-        sandbox.path(),
         locate_fixture("schemas").join("bins.toml"),
-        HashMap::from_iter([map_config_environment(HostOS::Windows, HostArch::X64)]),
+        |config| {
+            config.host(HostOS::Windows, HostArch::X64);
+        },
     );
 
     assert_eq!(
@@ -249,4 +260,176 @@ fn locates_windows_bin() {
             .exe_path,
         Some("win/moon.exe".into())
     );
+}
+
+mod primary {
+    use super::*;
+
+    #[test]
+    fn sets_primary_config() {
+        let sandbox = create_empty_proto_sandbox();
+        let plugin = sandbox.create_schema_plugin_with_config(
+            "schema-test",
+            locate_fixture("schemas").join("primary.toml"),
+            |config| {
+                config.host(HostOS::MacOS, HostArch::X64);
+            },
+        );
+
+        let config = plugin
+            .locate_executables(LocateExecutablesInput {
+                context: ToolContext {
+                    version: VersionSpec::parse("20.0.0").unwrap(),
+                    ..Default::default()
+                },
+            })
+            .primary
+            .unwrap();
+
+        assert_eq!(config.exe_path, Some("bin/moon".into()));
+        assert!(config.no_shim);
+        assert_eq!(
+            config.shim_before_args,
+            Some(StringOrVec::Vec(vec!["-v".into()]))
+        );
+    }
+
+    #[test]
+    fn auto_adds_exe_to_bin_on_windows() {
+        let sandbox = create_empty_proto_sandbox();
+        let plugin = sandbox.create_schema_plugin_with_config(
+            "schema-test",
+            locate_fixture("schemas").join("primary.toml"),
+            |config| {
+                config.host(HostOS::Windows, HostArch::X64);
+            },
+        );
+
+        assert_eq!(
+            plugin
+                .locate_executables(LocateExecutablesInput {
+                    context: ToolContext {
+                        version: VersionSpec::parse("20.0.0").unwrap(),
+                        ..Default::default()
+                    },
+                })
+                .primary
+                .unwrap()
+                .exe_path,
+            Some("bin/moon.exe".into())
+        );
+    }
+
+    #[test]
+    fn primary_path_doesnt_override_platform_path() {
+        let sandbox = create_empty_proto_sandbox();
+        let plugin = sandbox.create_schema_plugin_with_config(
+            "schema-test",
+            locate_fixture("schemas").join("primary-platform.toml"),
+            |config| {
+                config.host(HostOS::Linux, HostArch::X64);
+            },
+        );
+
+        assert_eq!(
+            plugin
+                .locate_executables(LocateExecutablesInput {
+                    context: ToolContext {
+                        version: VersionSpec::parse("20.0.0").unwrap(),
+                        ..Default::default()
+                    },
+                })
+                .primary
+                .unwrap()
+                .exe_path,
+            Some("lin/moon".into())
+        );
+    }
+}
+
+mod secondary {
+    use super::*;
+
+    #[test]
+    fn sets_secondary_config() {
+        let sandbox = create_empty_proto_sandbox();
+        let plugin = sandbox.create_schema_plugin_with_config(
+            "schema-test",
+            locate_fixture("schemas").join("secondary.toml"),
+            |config| {
+                config.host(HostOS::MacOS, HostArch::X64);
+            },
+        );
+
+        let secondary = plugin
+            .locate_executables(LocateExecutablesInput {
+                context: ToolContext {
+                    version: VersionSpec::parse("20.0.0").unwrap(),
+                    ..Default::default()
+                },
+            })
+            .secondary;
+
+        let foo = secondary.get("foo").unwrap();
+
+        assert_eq!(foo.exe_path, Some("bin/foo".into()));
+
+        let bar = secondary.get("bar").unwrap();
+
+        assert_eq!(bar.exe_path, Some("bin/bar".into()));
+        assert!(bar.no_bin);
+        assert_eq!(
+            bar.shim_env_vars,
+            Some(HashMap::from_iter([("BAR".into(), "bar".into())]))
+        );
+
+        let baz = secondary.get("baz").unwrap();
+
+        assert_eq!(baz.exe_path, Some("bin/baz".into()));
+        assert_eq!(baz.exe_link_path, Some("bin/baz-link".into()));
+        assert!(baz.no_shim);
+
+        let qux = secondary.get("qux").unwrap();
+
+        assert_eq!(qux.exe_path, Some("bin/qux.js".into()));
+        assert_eq!(qux.parent_exe_name, Some("node".into()));
+    }
+
+    #[test]
+    fn auto_adds_exe_to_bin_on_windows() {
+        let sandbox = create_empty_proto_sandbox();
+        let plugin = sandbox.create_schema_plugin_with_config(
+            "schema-test",
+            locate_fixture("schemas").join("secondary.toml"),
+            |config| {
+                config.host(HostOS::Windows, HostArch::X64);
+            },
+        );
+
+        let secondary = plugin
+            .locate_executables(LocateExecutablesInput {
+                context: ToolContext {
+                    version: VersionSpec::parse("20.0.0").unwrap(),
+                    ..Default::default()
+                },
+            })
+            .secondary;
+
+        assert_eq!(
+            secondary.get("foo").unwrap().exe_path,
+            Some("bin/foo.exe".into())
+        );
+        assert_eq!(
+            secondary.get("bar").unwrap().exe_path,
+            Some("bin/bar.exe".into())
+        );
+        assert_eq!(
+            secondary.get("baz").unwrap().exe_path,
+            Some("bin/baz.exe".into())
+        );
+        assert_eq!(
+            secondary.get("qux").unwrap().exe_path,
+            Some("bin/qux.js".into())
+        );
+    }
 }